Moving status icon related code into status_icon folder
