+++ components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_service/interactive/status_icon/BUILD.gn
@@ -1,0 +1,25 @@
+# Copyright (c) 2023 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+source_set("status_icon") {
+  sources = [
+    "brave_vpn_menu_model.cc",
+    "brave_vpn_menu_model.h",
+    "native_popup_menu.cc",
+    "native_popup_menu.h",
+    "status_icon.cc",
+    "status_icon.h",
+    "status_tray.cc",
+    "status_tray.h",
+  ]
+
+  public_deps = [ "//skia" ]
+
+  deps = [
+    "//brave/components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_service/interactive/common",
+    "//ui/base",
+    "//ui/gfx",
+  ]
+}
